conflictssss

+++ src/services/chatbot/chatbot_service.py
@@ -139,7 +139,7 @@ class DementiaChatbot:
             # Load base model
             base_model = AutoModelForCausalLM.from_pretrained(
                 self.base_model_name,
-                torch_dtype=torch.float16 if self.device in ["cuda", "mps"] else torch.float32,
+                torch_dtype=torch.float16 if self.device in ["cuda", "mps"] else torch.bfloat16,
                 device_map="auto" if self.device == "cuda" else None,
                 trust_remote_code=True,
                 token=hf_token  # Use token for private models
